requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
drf-yasg>=1.21
python-decouple>=3.6
Pillow>=9.0
